pyproject.toml
[tool.poetry]
name = "fastapi-resource-server"
version = "0.1.0"
description = "Build resource servers with FastAPI"
authors = ["Livio Ribeiro <livio@codata.pb.gov.br>"]
packages = [
    { include = "*", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.6.2"
fastapi = "^0"
python-jose = "^3.2"

[tool.poetry.dev-dependencies]
pytest = "^6.2"
pytest-cov = "^2.12.0"
pytest-mock = "^3.6.1"
black = "^21.5b1"
flake8 = "^3.9.2"
mypy = "^0.812"
isort = "^5.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
